Update working_controller.py
Working_controller.py is the development version of the PID controller. Main.py is the actual code housed on the Pico. This commit:  better values for the PID constants; default set point chosen to give measured time for 1 screw turn  79s. On board indicator LED added. Also, importantly, upper and lower values for PWM duty cycle modified. It appears the PWM cannot go all the way to 100% (65535) duty cycle so limit to 65000. At the lower limit, the motor is essentially stationary below 2000.